+++ src/main/java/com/haxwell/disposableIncomeScheduler/report/commands/SubtractCommand.java
@@ -2,34 +2,34 @@ package com.haxwell.disposableIncomeScheduler.report.commands;
 
 public class SubtractCommand extends Command {
 	
-	protected int amt = -1;
+	protected long amt = -1;
 	protected String grp = null;
 	protected String desc = null;
 	private String bufferString;
 	
-	public SubtractCommand(int amount, String desc) {
+	public SubtractCommand(long amount, String desc) {
 		this.amt = amount;
 		this.desc = desc; 
 	}
 
 	public SubtractCommand(double amount, String desc) {
-		this.amt = (int)Math.round(amount);
+		this.amt = (long)Math.round(amount);
 		this.desc = desc;
 	}
 	
-	public SubtractCommand(int amount, String desc, String group) {
+	public SubtractCommand(long amount, String desc, String group) {
 		this.amt = amount;
 		this.grp = group;
 		this.desc = desc;
 	}
 	
 	public SubtractCommand(double amount, String desc, String group) {
-		this.amt = (int)Math.round(amount);
+		this.amt = (long)Math.round(amount);
 		this.grp = group;
 		this.desc = desc;
 	}
 
-	public int getAmount() {
+	public long getAmount() {
 		return amt;
 	}
